✨Feat: Implement business logic, use cases and unit tests for user domain
- UserController 의 잔액 조회 API (`getBalance`), 잔액 조회 API (`getBalance`) 구현
- UserService 의 잔액 조회 (`getBalance`), 잔액 조회 (`getBalance`) 메서드 구현
- UserService 에 대한 단위 테스트 추가

Resolved #5, #9

+++ src/main/java/com/jane/ecommerce/interfaces/dto/user/ChargeRequest.java
@@ -10,5 +10,5 @@ import lombok.Setter;
 public class ChargeRequest {
 
     String userId;
-    int amount;
+    Long amount;
 }
