+++ src/main/java/com/jane/ecommerce/interfaces/dto/user/ChargeResponse.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class ChargeResponse {
-    int currentBalance;
+    Long currentBalance;
 }
